+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [`7.3.1.dev0` (unreleased)](https://github.com/kdeldycke/meta-package-manager/compare/v7.3.0...main)
+## [`7.4.0.dev0` (unreleased)](https://github.com/kdeldycke/meta-package-manager/compare/v7.3.0...main)
 
 > [!WARNING]
 > This version is **not released yet** and is under active development.